Addressing code review comments
